[dev]订单模块的调整

+++ order-server/order-server-api/src/main/java/com/example/order/rpc/IPayOrderRPC.java
@@ -43,8 +43,7 @@ public interface IPayOrderRPC {
      * @return
      */
     @PutMapping("/{id}")
-    RPCResult<Integer> updateByAccountId(@PathVariable("id") Integer id,
-                                         @RequestParam("payAmount") Integer payAmount);
+    RPCResult<Integer> updateByAccountId(@PathVariable("id") Long id, @RequestParam("payAmount") Integer payAmount);
 
     /**
      * 获取全部订单
